行为机的Root代码块中可以定义特定生命周期回调 eg. BeforeBehaviorReload

+++ Unity/Assets/Scripts/Loader/Box2D/b2GUIController.cs
@@ -204,7 +204,7 @@ namespace ET
                                 }
                                 
                                 ImGui.SliderFloat("Hertz", ref timelinePlayer.Hertz, 0f, 120f, "%.0f hz");
-                                EventSystem.Instance.Invoke(new UpdateHertzCallback() { instanceId = timelinePlayer.instanceId,Hertz = timelinePlayer.Hertz});
+                                EventSystem.Instance.Invoke(new UpdateHertzCallback() { instanceId = timelinePlayer.instanceId,Hertz = (int)timelinePlayer.Hertz});
                                 
                                 ImGui.TreePop();
                             }
